Add policy company owner

+++ app/Http/Controllers/CompanyUserController.php
@@ -15,17 +15,23 @@ class CompanyUserController extends Controller
 {
     public function index(Company $company): View
     {
+        $this->authorize('viewAny', $company);
+
         $users = $company->users()->get();
         return view('companies.users.index', compact('users', 'company'));
     }
 
     public function create(Company $company): View
     {
+        $this->authorize('create', $company);
+
         return view('companies.users.create', compact('company'));
     }
 
     public function store(Company $company, StoreUserRequest $request)
     {
+        $this->authorize('create', $company);
+
         $company->users()->create([
             'name' => $request->input('name'),
             'email' => $request->input('email'),
@@ -37,17 +43,23 @@ class CompanyUserController extends Controller
 
     public function edit(Company $company, User $user)
     {
+        $this->authorize('update', $company);
+
         return view('companies.users.edit', compact('company', 'user'));
     }
 
     public function update(UpdateUserRequest $request, Company $company, User $user)
     {
+        $this->authorize('update', $company);
+
         $user->update($request->validated());
         return to_route('companies.users.index', $company);
     }
 
     public function destroy(Company $company, User $user)
     {
+        $this->authorize('delete', $company);
+
         $user->delete();
         return to_route('companies.users.index', $company);
     }
